git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@2704 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/Multiple/File_Mpeg4_Elements.cpp
@@ -1465,7 +1465,7 @@ void File_Mpeg4::moov_meta_ilst_xxxx_data()
                             return;
                         case Elements::moov_meta__covr :
                             {
-                            std::string Data_Raw((const char*)(Buffer+(size_t)(Buffer_Offset+Element_Offset)), Element_Size-Element_Offset);
+                            std::string Data_Raw((const char*)(Buffer+(size_t)(Buffer_Offset+Element_Offset)), (size_t)(Element_Size-Element_Offset));
                             std::string Data_Base64(Base64::encode(Data_Raw));
                             Skip_XX(Element_Size-Element_Offset, "Data");
 
@@ -1523,7 +1523,7 @@ void File_Mpeg4::moov_meta_ilst_xxxx_data()
                     {
                         case Elements::moov_meta__covr :
                             {
-                            std::string Data_Raw((const char*)(Buffer+(size_t)(Buffer_Offset+Element_Offset)), Element_Size-Element_Offset);
+                            std::string Data_Raw((const char*)(Buffer+(size_t)(Buffer_Offset+Element_Offset)), (size_t)(Element_Size-Element_Offset));
                             std::string Data_Base64(Base64::encode(Data_Raw));
                             Skip_XX(Element_Size-Element_Offset, "Data");
